+++ packages/build-utils/src/index.ts
@@ -29,6 +29,7 @@ import {
   getEnvForPackageManager,
   getNodeVersion,
   getPathForPackageManager,
+  detectPackageManager,
   getSpawnOptions,
   getNodeBinPath,
   getNodeBinPaths,
@@ -74,6 +75,7 @@ export {
   getNodeBinPath,
   getNodeBinPaths,
   getSupportedNodeVersion,
+  detectPackageManager,
   runNpmInstall,
   runBundleInstall,
   runPipInstall,
@@ -111,6 +113,10 @@ export * from './schemas';
 export * from './types';
 export * from './errors';
 
+export * from './trace';
+
 export { NODE_VERSIONS } from './fs/node-version';
 
 export { getInstalledPackageVersion } from './get-installed-package-version';
+
+export { defaultCachePathGlob } from './default-cache-path-glob';