+++ packages/cli/src/commands/integration/add.ts
@@ -23,7 +23,6 @@ import { fetchInstallations } from '../../util/integration/fetch-installations';
 import { fetchIntegration } from '../../util/integration/fetch-integration';
 import output from '../../output-manager';
 import { IntegrationAddTelemetryClient } from '../../util/telemetry/commands/integration/add';
-import confirm from '../../util/input/confirm';
 
 export async function add(client: Client, args: string[]) {
   const telemetry = new IntegrationAddTelemetryClient({
@@ -115,22 +114,24 @@ export async function add(client: Client, args: string[]) {
   const metadataSchema = product.metadataSchema;
   const metadataWizard = createMetadataWizard(metadataSchema);
 
-  // At the time of writing, we don't support native integrations besides storage products.
+  // At the time of writing, we don't support native integrations besides storage & video products.
   // However, when we introduce new categories, we avoid breaking this version of the CLI by linking all
   // non-storage categories to the dashboard.
   // product.type is the old way of defining categories, while the protocols are the new way.
   const isPreProtocolStorageProduct = product.type === 'storage';
   const isPostProtocolStorageProduct =
     product.protocols?.storage?.status === 'enabled';
+  const isVideoProduct = product.protocols?.video?.status;
   const isStorageProduct =
     isPreProtocolStorageProduct || isPostProtocolStorageProduct;
+  const isSupportedProductType = isStorageProduct || isVideoProduct;
 
   // The provisioning via cli is possible when
   // 1. The integration was installed once (terms have been accepted)
   // 2. The provider-defined metadata is supported (does not use metadata expressions etc.)
   // 3. The product type is supported
   const provisionResourceViaCLIIsSupported =
-    installation && metadataWizard.isSupported && isStorageProduct;
+    installation && metadataWizard.isSupported && isSupportedProductType;
 
   if (!provisionResourceViaCLIIsSupported) {
     const projectLink = await getOptionalLinkedProject(client);
@@ -139,8 +140,7 @@ export async function add(client: Client, args: string[]) {
       return projectLink.exitCode;
     }
 
-    const openInWeb = await confirm(
-      client,
+    const openInWeb = await client.input.confirm(
       !installation
         ? 'Terms have not been accepted. Open Vercel Dashboard?'
         : 'This resource must be provisioned through the Web UI. Open Vercel Dashboard?',
@@ -175,8 +175,7 @@ async function getOptionalLinkedProject(client: Client) {
     return;
   }
 
-  const shouldLinkToProject = await confirm(
-    client,
+  const shouldLinkToProject = await client.input.confirm(
     'Do you want to link this resource to the current project?',
     true
   );
@@ -375,7 +374,7 @@ async function confirmProductSelection(
     `${chalk.dim(`- ${chalk.bold('Plan:')} ${billingPlan.name}`)}\n`
   );
 
-  return confirm(client, 'Confirm selection?', true);
+  return client.input.confirm('Confirm selection?', true);
 }
 
 async function provisionStorageProduct(