+++ packages/cli/src/commands/project/rm.ts
@@ -3,7 +3,6 @@ import ms from 'ms';
 import type Client from '../../util/client';
 import { emoji, prependEmoji } from '../../util/emoji';
 import { isAPIError } from '../../util/errors-ts';
-import confirm from '../../util/input/confirm';
 import { getCommandName } from '../../util/pkg-name';
 import { ProjectRmTelemetryClient } from '../../util/telemetry/commands/project/rm';
 import output from '../../output-manager';
@@ -87,5 +86,8 @@ async function readConfirmation(
     )
   );
 
-  return await confirm(client, `${chalk.bold.red('Are you sure?')}`, false);
+  return await client.input.confirm(
+    `${chalk.bold.red('Are you sure?')}`,
+    false
+  );
 }