+++ packages/cli/src/util/domains/purchase-domain-if-available.ts
@@ -5,7 +5,6 @@ import type Client from '../client';
 import eraseLines from '../output/erase-lines';
 import getDomainPrice from './get-domain-price';
 import getDomainStatus from './get-domain-status';
-import confirm from '../input/confirm';
 import purchaseDomain from './purchase-domain';
 import stamp from '../output/stamp';
 import * as ERRORS from '../errors-ts';
@@ -50,8 +49,7 @@ export default async function purchaseDomainIfAvailable(
     );
 
     if (
-      !(await confirm(
-        client,
+      !(await client.input.confirm(
         `Buy ${chalk.underline(domain)} for ${chalk.bold(
           `$${price}`
         )} (${plural('yr', period, true)})?`,